+++ tests/Application.UnitTests/Common/Behaviors/LoggingBehaviorTests.cs
@@ -1,6 +1,6 @@
 ﻿using Application.Common.Behaviors;
 using Application.Common.Interfaces;
-using Application.UnitTests.Helpers;
+using Application.UnitTests.TestHelpers;
 using Microsoft.Extensions.Logging;
 using Moq;
