+++ tests/Application.UnitTests/Common/Behaviors/PerformanceBehaviorTests.cs
@@ -1,6 +1,6 @@
 ﻿using Application.Common.Behaviors;
 using Application.Common.Interfaces;
-using Application.UnitTests.Helpers;
+using Application.UnitTests.TestHelpers;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Moq;
