+++ tests/Application.UnitTests/Common/Behaviors/UnhandledExceptionBehaviorTests.cs
@@ -1,5 +1,5 @@
 ﻿using Application.Common.Behaviors;
-using Application.UnitTests.Helpers;
+using Application.UnitTests.TestHelpers;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Moq;
