+++ tests/Application.UnitTests/Common/Behaviors/ValidationBehaviorTests.cs
@@ -1,5 +1,5 @@
 ﻿using Application.Common.Behaviors;
-using Application.UnitTests.Helpers;
+using Application.UnitTests.TestHelpers;
 using FluentValidation;
 using FluentValidation.Results;
 using MediatR;
